[ios, macos] Dry run mode
The output of this mode isn’t a good indicator of whether any files would’ve needed to be changed, because the presence of a conditional compilation block in one of the test methods means this script would always change and revert the corresponding comment.

+++ platform/darwin/scripts/update-examples.js
@@ -128,7 +128,9 @@ function completeExamples(os) {
         console.log('Skipping', path);
       } else {
         console.log('Updating', path);
-        fs.writeFileSync(path, newSrc);
+        if (['0', 'false'].indexOf(process.env.DRY_RUN || '0') !== -1) {
+          fs.writeFileSync(path, newSrc);
+        }
       }
     });
   });
